+++ mq/main/mq-broker/broker-core/src/main/java/com/sun/messaging/jmq/jmsserver/service/MetricManager.java
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2000, 2017 Oracle and/or its affiliates. All rights reserved.
- * Copyright (c) 2022 Contributors to Eclipse Foundation. All rights reserved.
+ * Copyright (c) 2022, 2025 Contributors to Eclipse Foundation. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0, which is available at
@@ -226,7 +226,7 @@ public class MetricManager implements ConfigListener {
 
         long currentTime = System.currentTimeMillis();
         // long elapsedSecs = (currentTime - lastSampleTime) / 1000;
-        float elapsedSecs = (currentTime - lastSampleTime) / (float) 1000;
+        float elapsedSecs = (currentTime - lastSampleTime) / 1000.0f;
 
         MetricCounters totals = getMetricCounters(null);
